separate Trie and Node classes
