+++ Public/Src/Sandbox/MacOs/Sandbox/Src/Utilities/Trie.cpp
@@ -6,123 +6,6 @@
 
 #define super OSObject
 
-// ================================== class Node ==================================
-
-OSDefineMetaClassAndStructors(Node, OSObject)
-
-uint Node::s_numUintNodes = 0;
-uint Node::s_numPathNodes = 0;
-
-Node* Node::create(uint maxKey, uint key)
-{
-    Node *instance = new Node;
-    if (instance != nullptr)
-    {
-        if (maxKey == s_uintNodeMaxKey)      OSIncrementAtomic(&s_numUintNodes);
-        else if (maxKey == s_pathNodeMaxKey) OSIncrementAtomic(&s_numPathNodes);
-
-        if (!instance->init(maxKey, key))
-        {
-            OSSafeReleaseNULL(instance);
-        }
-    }
-
-    return instance;
-}
-
-bool Node::init(uint maxKey, uint key)
-{
-    if (!super::init())
-    {
-        return false;
-    }
-
-    assert(key < maxKey);
-
-    key_    = key;
-    maxKey_ = maxKey;
-    record_ = nullptr;
-
-    next_     = nullptr;
-    children_ = nullptr;
-
-    lock_ = IORecursiveLockAlloc();
-    if (!lock_)
-    {
-        return false;
-    }
-
-    return true;
-}
-
-void Node::free()
-{
-    // intentionally not calling OSSafeReleaseNULL on children_ and next_
-    // because Trie is responsible for releasing all its nodes
-    next_ = nullptr;
-    children_ = nullptr;
-
-    OSSafeReleaseNULL(record_);
-    IORecursiveLockFree(lock_);
-
-    if (maxKey_ == s_uintNodeMaxKey)      OSDecrementAtomic(&s_numUintNodes);
-    else if (maxKey_ == s_pathNodeMaxKey) OSDecrementAtomic(&s_numPathNodes);
-
-    super::free();
-}
-
-Node* Node::findChild(uint key, bool createIfMissing, IORecursiveLock *lock)
-{
-    if (key < 0 || key >= maxKey_)
-    {
-        return nullptr;
-    }
-
-    Monitor __monitor(lock); // this will only acquire the lock if lock is not null
-
-    Node *prev = nullptr;
-    Node *curr = children_;
-    while (curr != nullptr && curr->key_ != key)
-    {
-        prev = curr;
-        curr = curr->next_;
-    }
-
-    if (curr != nullptr)
-    {
-        // found it
-        assert(curr->key_ == key);
-        return curr;
-    }
-    else if (!createIfMissing)
-    {
-        // didn't find it and shouldn't create it
-        return nullptr;
-    }
-    else if (lock == nullptr)
-    {
-        // didn't find it and didn't acquire lock --> must do it all over again with a lock
-        return findChild(key, createIfMissing, lock_);
-    }
-    else
-    {
-        // didn't find it and we are holding the lock -> create a new node and link it
-        Node *newNode = Node::create(maxKey_, key);
-        if (prev != nullptr)
-        {
-            prev->next_ = newNode;
-        }
-        else
-        {
-            assert(children_ == nullptr);
-            children_ = newNode;
-        }
-        return newNode;
-    }
-}
-
-// ================================== class Trie ==================================
-
 OSDefineMetaClassAndStructors(Trie, OSObject)
 
 Trie* Trie::create(TrieKind kind)
@@ -161,7 +44,7 @@ bool Trie::init(TrieKind kind)
 
 void Trie::free()
 {
-    traverse(/*computeKey*/ false, /*callbackArgs*/ nullptr, [](Trie*, void*, uint64_t, Node *n)
+    traverse(/*computeKey*/ false, /*callbackArgs*/ nullptr, [](void*, uint64_t, Node *n)
              {
                  OSSafeReleaseNULL(n);
              });
@@ -632,7 +515,7 @@ void Trie::forEach(void *callbackArgs, for_each_fn callback)
 {
     typedef struct { for_each_fn callback; void *args; } State;
     State state = { .callback = callback, .args = callbackArgs };
-    traverse(/*computeKey*/ kind_ == kUintTrie, /*callbackArgs*/ &state, [](Trie *me, void *s, uint64_t key, Node *node)
+    traverse(/*computeKey*/ kind_ == kUintTrie, /*callbackArgs*/ &state, [](void *s, uint64_t key, Node *node)
              {
                  State *state = (State*)s;
                  OSObject *record = node->record_;
@@ -647,9 +530,9 @@ void Trie::forEach(void *callbackArgs, for_each_fn callback)
 
 void Trie::removeMatching(void *filterArgs, filter_fn filter)
 {
-    typedef struct { filter_fn filter; void *args; } State;
-    State state = { .filter = filter, .args = filterArgs };
-    traverse(/*computeKey*/ false, /*callbackArgs*/ &state, [](Trie *me, void *s, uint64_t, Node *node)
+    typedef struct { filter_fn filter; void *args; Trie *me; } State;
+    State state = { .filter = filter, .args = filterArgs, .me = this };
+    traverse(/*computeKey*/ false, /*callbackArgs*/ &state, [](void *s, uint64_t, Node *node)
              {
                  State *state = (State*)s;
                  OSObject *record = node->record_;
@@ -658,87 +541,14 @@ void Trie::removeMatching(void *filterArgs, filter_fn filter)
                      record->retain();
                      if (state->filter(state->args, record))
                      {
-                         me->remove(node);
+                         state->me->remove(node);
                      }
                      record->release();
                  }
              });
 }
 
-typedef struct Stack {
-    Node *node;
-    Stack *next;
-    uint32_t depth;
-    uint64_t key;
-} Stack;
-
-static void push(Stack **stack, Node *node, uint64_t path, uint32_t depth)
-{
-    if (node == nullptr) return;
-
-    Stack *top = IONew(Stack, 1);
-    top->node  = node;
-    top->next  = *stack;
-    top->key   = path;
-    top->depth = depth;
-
-    *stack = top;
-}
-
-static Node* pop(Stack **stack)
-{
-    Stack *top = *stack;
-    *stack = top->next;
-    Node *node = top->node;
-    IODelete(top, Stack, 1);
-    return node;
-}
-
-static uint64_t s_pow10[] =
-{
-    1,
-    10,
-    100,
-    1000,
-    10000,
-    100000,
-    1000000,
-    10000000,
-    100000000,
-    1000000000,
-    10000000000,
-    100000000000,
-    1000000000000
-};
-
-static int s_powLen = sizeof(s_pow10)/sizeof(s_pow10[0]);
-
-static uint64_t pow10(int exp)
-{
-    if (exp < s_powLen) return s_pow10[exp];
-    uint64_t result = 1;
-    while (--exp >= 0) result *= 10;
-    return result;
-}
-
 void Trie::traverse(bool computeKey, void *callbackArgs, traverse_fn callback)
 {
-    Stack *stack = nullptr;
-    push(&stack, root_, /*key*/ 0, /*depth*/ 0);
-    while (stack != nullptr)
-    {
-        uint64_t key = stack->key;
-        uint32_t depth = stack->depth;
-
-        Node *toVisit = pop(&stack);
-        Node *curr = toVisit->children_;
-        while (curr)
-        {
-            push(&stack, curr, computeKey ? (curr->key_ * pow10(depth) + key) : 0, depth + 1);
-            curr = curr->next_;
-        }
-
-        // the callback may deallocate 'curr' node, hence this must be the last statement in this loop
-        callback(this, callbackArgs, key, toVisit);
-    }
+    root_->traverse(computeKey, callbackArgs, callback);
 }
